requirements.txt
discord.py
pillow
aiohttp
python-dotenv
cryptography